address a bunch of lints
