+++ src/core/transposer/engine/buffered_item.rs
@@ -7,51 +7,16 @@ use std::{
     task::{Context, Poll},
 };
 
-use crate::core::Transposer;
+use crate::core::{
+    transposer::engine::update_item::{EventsEmitted, UpdateItemData},
+    Transposer,
+};
 
 use super::{
     engine_time::EngineTime, input_buffer::InputBuffer, lazy_state::LazyState,
     transposer_frame::TransposerFrame, transposer_update::TransposerUpdate,
+    update_item::UpdateItem,
 };
-
-// pointer needs to be the top argument as its target may have pointers into inputs or transposer.
-#[pin_project]
-pub struct UpdateItem<'a, T: Transposer> {
-    #[pin]
-    pub time: EngineTime<'a, T::Time>,
-    // TODO: EngineTime and UpdateItemData both track the same thing. they probably should be merged.
-    pub data: UpdateItemData<T>,
-    pub events_emitted: EventsEmitted,
-}
-
-pub enum UpdateItemData<T: Transposer> {
-    Init(Box<T>),
-    Input(Box<[T::Input]>),
-    Schedule,
-}
-
-pub enum EventsEmitted {
-    Some,
-    None,
-    Pending,
-}
-
-impl EventsEmitted {
-    pub fn any(&self) -> bool {
-        match self {
-            Self::Some => true,
-            Self::None => false,
-            Self::Pending => false,
-        }
-    }
-    pub fn done(&self) -> bool {
-        match self {
-            Self::Some => true,
-            Self::None => true,
-            Self::Pending => false,
-        }
-    }
-}
 
 #[pin_project(project=BufferedItemProject)]
 pub struct BufferedItem<'a, T: Transposer> {
@@ -117,8 +82,7 @@ impl<'a, T: Transposer> BufferedItem<'a, T> {
             let update_future = Pin::into_inner_unchecked(update_future);
             *update_future = MaybeUninit::new(TransposerUpdate::new());
             let update_future = update_future.assume_init_mut();
-            let update_future = Pin::new_unchecked(update_future);
-            update_future
+            Pin::new_unchecked(update_future)
         };
 
         let transposer_frame: *mut _ = this.transposer_frame;
